refactor(html): make `html` as middleware
