+++ src/middleware/jsx/README.md
@@ -100,6 +100,53 @@ const List = () => (
 )
 ```
 
+## With html Middleware
+
+It's powerful to use JSX middleware with html middleware.
+For more information, see [html middleware docs](https://github.com/honojs/hono/tree/master/src/middleware/html).
+
+```tsx
+import { Hono } from 'hono'
+import { html } from 'hono/html'
+import { jsx } from 'hono/jsx'
+
+const app = new Hono()
+
+interface SiteData {
+  title: string
+  children?: any
+}
+
+const Layout = (props: SiteData) => html`<!DOCTYPE html>
+  <html>
+    <head>
+      <title>${props.title}</title>
+    </head>
+    <body>
+      ${props.children}
+    </body>
+  </html>`
+
+const Content = (props: { siteData: SiteData; name: string }) => (
+  <Layout {...props.siteData}>
+    <h1>Hello {props.name}</h1>
+  </Layout>
+)
+
+app.get('/:name', (c) => {
+  const { name } = c.req.param()
+  const props = {
+    name: name,
+    siteData: {
+      title: 'JSX with html sample',
+    },
+  }
+  return c.html(<Content {...props} />)
+})
+
+app.fire()
+```
+
 ## Tips for Cloudflare Workers
 
 It's useful to use Miniflare's`live-reload` option for developing.
